+++ src/main/java/barch/mc_extended/Armor/SapphireArmorMaterial.java
@@ -35,7 +35,7 @@ public class SapphireArmorMaterial implements ArmorMaterial {
 
     @Override
     public int getEnchantability() {
-        return 10;
+        return 15;
     }
 
     @Override