+++ src/main/java/barch/mc_extended/Armor/TinArmorMaterial.java
@@ -33,7 +33,7 @@ public class TinArmorMaterial implements ArmorMaterial {
 
     @Override
     public int getEnchantability() {
-        return 1;
+        return 5;
     }
 
     @Override