+++ src/main/resources/changelog.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-8-15
+
+### Changed
+- copper gear now has 25 enchantability
+- tin gear now has 5 enchantability
+- bronze gear now has 15 enchantability
+- ruby gear now has 15 enchantability
+- sapphire gear now has 15 enchantability
+- silver gear now has 20 enchantability
+- corundum gear now has 20 enchantability
+- bronze tools now are mining level 2
+
 ## [0.2.5] - 2023-8-4
 ### Added
 - ender cube
@@ -271,6 +283,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ruby block
 
 [0.2.5]: https://github.com/BarchamMal/MC-Extended/commit/
+[0.3.0]: https://github.com/BarchamMal/MC-Extended/commit/
+[0.2.5]: https://github.com/BarchamMal/MC-Extended/commit/97641cc3b4809d947953fd8fa6c8d16cbb0ba7e4
 [0.2.4]: https://github.com/BarchamMal/MC-Extended/commit/c7b83a5135e549532c4dd4b6657f140957cfd31f
 [0.2.3]: https://github.com/BarchamMal/MC-Extended/commit/747cc0ceb81f4288abf79e40784574a3411c76ac
 [0.2.2]: https://github.com/BarchamMal/MC-Extended/commit/b252f3e3333e4fd547f65bb483a0c393884aa9ca